Add missing CPU_* macros to android, and sync with Linux
Linux had the full set of CPU_* macros, but Android missed a few. Add
the missing ones: CPU_ALLOC_SIZE, CPU_COUNT_S, and CPU_COUNT.

Sync the definitions more closely with the Linux versions, so that
they're easier to keep in sync. The only differences now are the field
name: bits for Linux vs __bits for Android.